Allow arrays and hashes in signatures
TryFromSV is a new trait that allows conversion to fail with an explicit
desciptive message. For now it is only used by the new context method,
st_try_fetch, but not by arrays or hashes themselves. XS macro is
adjusted to use st_try_fetch when pulling arguments from the stack, and
panic with a new message in the case of error.