Add croak macro.
Panics were propagated to perl as exceptions, which made `panic!`
already behave a lot like croak behaves in XS, except for the panic
message printed by the default panic hook. The message is not necessary
when program handles the exception gracefully, and just duplicate if
not (itself perl will print unhandled exceptions).

This does not replace the panic hook, so panics that occur inside the
Rust libraries will still invoke the default (or user supplied) panic
hook.

Fixes #10.