Allow subroutines to return values with 'return'.
Before this commit, subroutines were declared as unit functions, and
were required to use xs_return! macro to push their return values on the
perl stack.

With this commit, a subroutine may return any value and have it
automatically pushed on the interpreter stack, as long as the value
implements Stackable trait. Stackable is implemented for IntoSV types,
and tuples of such types, allowing simple values and short lists to be
returned easily.

Calling st_prepush() and st_putback() when returning values is no longer
necessary, internal stack pointer is handled automatically by the
Context. Furthermore, st_prepush() was removed entirely, as the stack
pointer is set to the right position automatically during
initialization.

xs_return! macro is now unnecessary and was removed.