Change subcrate license to MIT/APACHE-2.0
Many corporate opensource usage policies require an MIT or APACHE 2.0 License. Other licenses
may be quite permissive, but may lack approval from the requisite legal departments as to enable
their ease of use. This is why the standard license for the rust ecosystem is a MIT/APACHE-2.0
dual license.

As of the time of this writing, and for all relevant preceeding code, Daniel Norman is the sole
committer and author of the code within `perlxs_derive` `perlxs_derive_internals` (hereafter known
as `crates/macro-derive` and `crates/macro-support`, respectively)

Daniel Norman hereby approves of the change of license for the aforementioned subcrates to be used
under the terms of either the MIT License *OR* APACHE-2.0 License, as chosen by the user.