Fix compilation bug
perlxs_derive included dependency on perl_xs, which, on later versions
of rustc caused the compiled proc macro library to require perl's
symbols. Since the proc macro is loaded by the rustc and not perl,
these symbols could not be resolved and compilation failed.

Fix the problem by removing the unnecessary import.